% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMeansEqual.R
\name{kMeansEqual}
\alias{kMeansEqual}
\title{Equal-size kmeans}
\usage{
kMeansEqual(
  kdat,
  k,
  centers = NULL,
  add.centers = FALSE,
  assigned.to = "assigned"
)
}
\arguments{
\item{kdat}{A data frame with 2 columns.}

\item{k}{The number of clusters. Each of the `k` clusters will contain at least `floor(nrow(kdat)/k)` elements.}

\item{centers}{An optional data frame with `k` rows and 2 columns indicating the centers of each of the `k` clusters. If `NULL` (default), the cluster centers will be calculated with the `kmeans` function.}

\item{add.centers}{A boolean indicating wheter new cluster centers need to be calculated. 
If `TRUE`, an attribute named `"centers"`  will be added to the output that contains the data frame with the new cluster centers. 
This attribute can immediately be plugged into the `centers` argument if this function is to be executed iteratively. Defaults to `FALSE`.}

\item{assigned.to}{A character vector indicating the name of the column that will contain the cluster identifier. Defaults to `"assigned"`.}
}
\description{
`kMeansEqual` Assigns `k` cluster identifiers to a data frame `kdat` whereby each cluster is of (nearly) equal size.
}
